syncQueue -> queue, iterateSyncQueue -> processQueue
