+++ src/clone-deep/clone-deep-utils/handle-customizer.js
@@ -7,7 +7,7 @@ import { getWarning, Warning } from '../../utils/clone-deep-warning.js';
  * necessary. Errors from the customizers are also handled here.
  * @param {Object} spec
  * @param {import('../../types').Log} spec.log
- * @param {import('../../types').SyncQueueItem[]} spec.syncQueue
+ * @param {import('../../types').QueueItem[]} spec.queue
  * @param {import('../../types').Customizer} spec.customizer
  * @param {any} spec.value
  * @param {(clone: any) => any} spec.saveClone
@@ -21,7 +21,7 @@ import { getWarning, Warning } from '../../utils/clone-deep-warning.js';
  * }}
  */
 export const handleCustomizer = (
-    { log, syncQueue, customizer, value, saveClone, doThrow }) => {
+    { log, queue, customizer, value, saveClone, doThrow }) => {
 
     /** @type {any} */
     let cloned;
@@ -62,7 +62,7 @@ export const handleCustomizer = (
                     additionalValues.forEach((object) => {
                         if (typeof object === 'object'
                             && typeof object.assigner === 'function') {
-                            syncQueue.push({
+                            queue.push({
                                 value: object.value,
                                 parentOrAssigner: object.assigner
                             });
